+++ CHANGELOG.md
@@ -1,9 +1,14 @@
 # vNext
 
 - Switch from deprecated v3 Gitlab API to v4. Thanks @SfinxNT (!20)
+- Add support of .netrc file to get personal access token, under new `--netrc|-n` option. Allows to use multiple Gitlab instances with auth in a same dev environment
+- Guess project path, and use it in the API URL.
+- Added a `--project-path` option
 - Added a `--merged-yaml` option to allow merged yaml to be returned in response from gitlab API
+- Improve Gitlab API HTTP error code handling
 - Upgrade to Go 1.23
 - Use Goreleaser
+- Bugfixes
 
 # v2.2.0